Server/requirements.txt
Django==4.0
django-cors-headers
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-filter
matplotlib==3.7.2
pandas
numpy
google-auth==2.21.0
requests==2.25.1
Pillow